feat: 部署文档 + GitHub Actions CI + API 状态页 + 修复 PostgreSQL 二次启动 panic
- 新增 deploy.md 部署指南（PostgreSQL/Nginx 图床子域名/systemd）
- 新增 GitHub Actions release workflow（musl 静态编译 + UPX 压缩）
- 新增后台 API 状态页面，展示公开端点列表
- 修复 PostgreSQL 下 ensure_root_user 类型不匹配导致二次启动 panic

+++ src/main.rs
@@ -336,7 +336,7 @@ async fn ensure_root_user(pool: &DbPool) {
             row.is_some()
         }
         DbPool::Postgres(pool) => {
-            let row: Option<(i64,)> =
+            let row: Option<(i32,)> =
                 sqlx::query_as("SELECT id FROM users WHERE username = 'root'")
                     .fetch_optional(pool)
                     .await
